done with Ajax for post and comment and with image uploadergit add -A .

+++ app/controllers/comments_controller.rb
@@ -7,34 +7,22 @@ class CommentsController < ApplicationController
     @post = Post.includes(:comments).find_by(id: params[:post_id])
     @comments = @post.comments.order("created_at DESC")
     @comment = Comment.new
-    authorize @comment
   end
 
   def show
     @comment = Comment.find_by(id: params[:id])
   end
 
-  # def new
-  #   @topic = Topic.find_by(id: params[:topic_id])
-  #   @post = @topic.posts.find_by(id: params[:post_id])
-  #   @comment = Comment.new
-  #   authorize @comment
-  # end
-
   def create
-
     @topic = Topic.find_by(id: params[:topic_id])
     @post = @topic.posts.find_by(id: params[:post_id])
-    # @comment = @post.comments.new(comment_params)
     @comment = current_user.comments.build(comment_params.merge(post_id: params[:post_id]))
     @new_comment=Comment.new
     authorize @comment
     if @comment.save
       flash.now[:success] = "You've created a new comment."
-      redirect_to topic_post_comments_path(@topic, @post)
     else
       flash.now[:danger] = @comment.errors.full_messages
-      redirect_to topic_post_comments_path(@topic, @post)
     end
   end
 
@@ -63,8 +51,11 @@ class CommentsController < ApplicationController
     @topic = Topic.find_by(id: params[:topic_id])
     @post = Post.find_by(id: params[:topic_id])
     @comment = Comment.find_by(id: params[:id])
+    authorize @comment
     if @comment.destroy
-      redirect_to topic_post_comments_path(@topic, @post)
+      flash.now[:success] = "You've deleted a new comment."
+    else
+      flash.now[:danger] = @comment.errors.full_messages
     end
 
   end
